requirements.txt
pandas
numpy
seaborn
scikit-learn
xgboost
plotly[express]
requests
python-dotenv